WIP: split Channel state and data models and complete it
